MOBI-1121 modify example to not dynamically change any input attributes

+++ icemobile/tags/icemobile-1.3.1_P03/icemobile/samples/jsf/mobileshowcase/src/main/java/org/icemobile/samples/mobileshowcase/view/examples/input/input/InputBean.java
@@ -70,7 +70,7 @@ public class InputBean extends ExampleImpl<InputBean> implements
     private String textArea;
     private String urlInput;
     private Double numberInput;
-    private Double numberStep = 1.0d;
+    private Float numberStep = 1.0f;
     private String dateInput;
 
 
@@ -150,11 +150,11 @@ public class InputBean extends ExampleImpl<InputBean> implements
         }
     }
 
-    public Double getNumberStep() {
+    public Float getNumberStep() {
         return numberStep;
     }
 
-    public void setNumberStep(Double numberStep) {
+    public void setNumberStep(Float numberStep) {
         this.numberStep = numberStep;
     }
 }
